Refactoring into seperate routers
