+++ server/routes/searchRouter.test.ts
@@ -3,15 +3,13 @@ import request from 'supertest'
 import { appWithAllRoutes, user } from './testutils/appSetup'
 import AuditService, { Page } from '../services/auditService'
 import SearchService from '../services/searchService'
-import OrderService from '../services/orderService'
+import { basicGetTest, GetRequestFixture } from './index.test'
 
 jest.mock('../services/auditService')
 jest.mock('../services/searchService')
-jest.mock('../services/orderService')
 
 const auditService = new AuditService(null) as jest.Mocked<AuditService>
 const searchService = new SearchService() as jest.Mocked<SearchService>
-const orderService = new OrderService() as jest.Mocked<OrderService>
 
 let app: Express
 
@@ -20,7 +18,6 @@ beforeEach(() => {
     services: {
       auditService,
       searchService,
-      orderService,
     },
     userSupplier: () => user,
   })
@@ -31,25 +28,19 @@ afterEach(() => {
 })
 
 describe('Core page basic GET requests', () => {
-  it.each<GetRequestFixture>([['index page', '/', 'Electronic Monitoring Datastore', Page.START_PAGE]])(
+  it.each<GetRequestFixture>([['search page', '/search', 'Search for case details', Page.SEARCH_PAGE]])(
     'should render %s',
     (pageName, route, titleText, auditType) => basicGetTest(pageName, route, titleText, auditType),
   )
 })
 
-export type GetRequestFixture = [pageName: string, route: string, titleText: string, auditType: Page]
-
-export function basicGetTest(pageName: string, route: string, titleText: string, auditType: Page) {
-  auditService.logPageView.mockResolvedValue(null)
-
-  return request(app)
-    .get(route)
-    .expect('Content-Type', /html/)
-    .expect(res => {
-      expect(res.text).toContain(titleText)
-      expect(auditService.logPageView).toHaveBeenCalledWith(auditType, {
-        who: user.username,
-        correlationId: expect.any(String),
+describe('Search results page', () => {
+  it('should call the SearchService to return data', () => {
+    return request(app)
+      .get('/search/results')
+      .expect('Content-Type', /html/)
+      .expect(res => {
+        expect(searchService.getOrders).toHaveBeenCalledTimes(1)
       })
-    })
-}
+  })
+})
